car-parts-sales-platform/backend/requirements.txt
Django>=5.2
djangorestframework>=3.15
mysqlclient>=2.2
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
django-filter>=23.5
Pillow>=10.0
python-dotenv>=1.0
passlib>=1.7
django-simple-history>=3.5
